create a static store for popups instead of a creator function
