+++ src/popup.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import createStore from './stores/createPopupStore';
+import popupStore from './stores/popupStore';
 import checkCurrentUser from './utils/checkCurrentUser';
 import './less/app.less';
-import MainList from './components/RenderMainList';
+import MainList from './components/MainList';
 
 const container = document.getElementById('wakatime');
-const store = createStore('WakaTime-Options');
+const store = popupStore;
 checkCurrentUser(store)(30 * 1000);
 
 ReactDOM.render(
